REDO_module_4/requirements.txt
Flask
pytest
pytest-cov
requests
beautifulsoup4
psycopg2-binary